move stuff out of modules directory that isn't a discreet testable function
