+++ builders/trailsBuilder.js
@@ -1,7 +1,7 @@
 const jsonfile = require('jsonfile')
 const helpers = require('@turf/helpers');
 const path = require('path').normalize;
-const gQuery = require('../modules/genericQuery');
+const gQuery = require('../db/genericQuery');
 const labelMaker = require('../modules/labelMaker').labelMaker;
 const exec = require('child_process').execSync;
 const fs = require('fs');
