+++ data/mapboxStaticData.js
